+++ src/Response.php
@@ -1,0 +1,16 @@
+<?php
+
+/**
+ * @package Harvest
+ * @license http://opensource.org/licenses/MIT
+ */
+
+declare(strict_types=1);
+
+namespace DecodeLabs\Harvest;
+
+use Psr\Http\Message\ResponseInterface as PsrResponse;
+
+interface Response extends PsrResponse
+{
+}