+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added nominal Response interface
+
 ## v0.2.3 (2023-11-08)
 * Fixed Exception stack handling in Dispatcher
 * Added permanence methods to Redirect responses